Fixes for stk mesh parallel transition.
1) unshare created nodes
2) unshare elements, make ghosted
3) stk_parallel_model set to STK P SM NO AURA
4) protect orphaned nodes

(From Sierra)

+++ packages/stk/stk_mesh/stk_mesh/base/BulkData.hpp
@@ -1139,7 +1139,7 @@ private:
                                        const PartVector & remove_parts); // Mod Mark
   void internal_adjust_entity_and_downward_connectivity_closure_count(stk::mesh::Entity entity,
                                                                       stk::mesh::Bucket *bucket_old,
-                                                                      uint16_t closureCountAdjustment); // Mod Mark
+                                                                      int closureCountAdjustment); // Mod Mark
 
   void internal_fill_new_part_list_and_removed_part_list(stk::mesh::Entity entity,
                                                            const PartVector & add_parts,
